Finished converting all the variables to their right data types so we can upload it to the SQL local database.
Dropped 'url' columns since it contains only links to Lending Club's website.
Changed the name of 'desc' column to 'loan_purpose' because the variable describes the purpose of the loan written by the borrower.
Converted 'title' column to a category data type and changed the name from 'title' to 'loan_title' because it sounds more intuitive and is the title given by the borrower.
Created a new variable 'zip_prefix' that contains the first 3 digits of the zip codes in the 'zip_code' column. I did this because a portion of the zip codes in the 'zip_code' column were blanked out with x's, presumably to maintain the privacy of the borrower.
Categorical vars like 'emp_title' (Employee title) and 'loan_title' still require cleaning because some categories have emojis and unnecessary symbols.