New chart added to track loan status overtime
I added new chart to show the percantage of loans that fully paid vs Charged Off between 2007 and 2018

Deleted code that led to errors while attempting to show the percantage of loans that fully paid vs Charged Off between 2007 and 2018.

Added titles/markdowns to show what each new code block is attempting to achieve to make it easier for stakeholders to read.